Improved executiong logging

+++ app/src/sagemaker_cleanup.py
@@ -155,8 +155,10 @@ class SageMakerCleanup:
                     delta = Helper.get_day_delta(resource_date)
 
                     if delta.days > ttl_days:
-                        if not self.settings.get("general", {}).get("dry_run", True):
-                            if resource_status == "InService":
+                        if resource_status == "InService":
+                            if not self.settings.get("general", {}).get(
+                                "dry_run", True
+                            ):
                                 try:
                                     self.client_sagemaker.stop_notebook_instance(
                                         NotebookInstanceName=resource_id,
@@ -168,13 +170,16 @@ class SageMakerCleanup:
                                     self.logging.error(sys.exc_info()[1])
                                     resource_action = "error"
                                     continue
-                                else:
-                                    self.logging.info(
-                                        f"SageMaker Notebook Instance '{resource_id}' was last modified {delta.days} days ago "
-                                        "and has been stopped."
-                                    )
-                                    resource_action = "stop"
-                            elif resource_status in ("Stopped", "Failed"):
+
+                            self.logging.info(
+                                f"SageMaker Notebook Instance '{resource_id}' was last modified {delta.days} days ago "
+                                "and has been stopped."
+                            )
+                            resource_action = "stop"
+                        elif resource_status in ("Stopped", "Failed"):
+                            if not self.settings.get("general", {}).get(
+                                "dry_run", True
+                            ):
                                 try:
                                     self.client_sagemaker.delete_notebook_instance(
                                         NotebookInstanceName=resource_id,
@@ -186,12 +191,12 @@ class SageMakerCleanup:
                                     self.logging.error(sys.exc_info()[1])
                                     resource_action = "error"
                                     continue
-                                else:
-                                    self.logging.info(
-                                        f"SageMaker Notebook Instance '{resource_id}' was last modified {delta.days} days ago "
-                                        "and has been deleted."
-                                    )
-                                    resource_action = "delete"
+
+                            self.logging.info(
+                                f"SageMaker Notebook Instance '{resource_id}' was last modified {delta.days} days ago "
+                                "and has been deleted."
+                            )
+                            resource_action = "delete"
                     else:
                         self.logging.debug(
                             f"SageMaker Notebook Instance '{resource_id}' was created {delta.days} days ago "
